+++ persistent-mysql/Database/Persist/MySQL.hs
@@ -45,11 +45,13 @@ import Control.Monad.Trans.Class (lift)
 import Control.Monad.Trans.Except (ExceptT, runExceptT)
 import Control.Monad.Trans.Reader (ReaderT, runReaderT)
 import Control.Monad.Trans.Writer (runWriterT)
+import Data.Proxy (Proxy(..))
 
 import Data.Acquire (Acquire, mkAcquire, with)
 import Data.Aeson
 import Data.Aeson.Types (modifyFailure)
 import Data.ByteString (ByteString)
+import qualified Data.ByteString.Lazy as BSL
 import Data.Conduit
 import qualified Data.Conduit.List as CL
 import Data.Either (partitionEithers)
@@ -58,6 +60,7 @@ import Data.Function (on)
 import Data.Int (Int64)
 import Data.List (find, groupBy, intercalate, sort)
 import qualified Data.List.NonEmpty as NEL
+import qualified Data.Map as Map
 import Data.Maybe (fromMaybe, listToMaybe, mapMaybe)
 import qualified Data.Map as Map
 import Data.Monoid ((<>))
@@ -335,6 +338,7 @@ getGetter field = go (MySQLBase.fieldType field)
       case m of
         Just g -> PersistLiteral g
         Nothing -> error "Unexpected null in database specific value"
+    go MySQLBase.Json       _ _  = convertPV PersistByteString
     -- Unsupported
     go other _ _ = error $ "MySQL.getGetter: type " ++
                       show other ++ " not supported."
@@ -453,7 +457,7 @@ addTable cols entity = AddTable $ concat
     , "("
     , idtxt
     , if null nonIdCols then [] else ","
-    , intercalate "," $ map showColumn nonIdCols
+    , intercalate "," $ map showCreateColumn nonIdCols
     , ")"
     ]
   where
@@ -467,7 +471,9 @@ addTable cols entity = AddTable $ concat
                 concat
                     [ " PRIMARY KEY ("
                     , intercalate ","
-                  $ map (escapeF . fieldDB) $ NEL.toList $ compositeFields pdef
+                        $ map (escapeF . fieldDB)
+                        $ NEL.toList
+                        $ compositeFields pdef
                     , ")"
                     ]
             EntityIdField idField ->
@@ -961,8 +967,14 @@ findAlters edef allDefs col@(Column name isNull type_ def gen _defConstraintName
 
 -- | Prints the part of a @CREATE TABLE@ statement about a given
 -- column.
-showColumn :: Column -> String
-showColumn (Column n nu t def gen _defConstraintName maxLen ref) = concat
+showCreateColumn :: Column -> String
+showCreateColumn = showColumn False
+
+showAlterColumn :: Column -> String
+showAlterColumn = showColumn True
+
+showColumn :: Bool -> Column -> String
+showColumn showReferences (Column n nu t def gen _defConstraintName maxLen ref) = concat
     [ escapeF n
     , " "
     , showSqlType t maxLen True
@@ -975,13 +987,21 @@ showColumn (Column n nu t def gen _defConstraintName maxLen ref) = concat
     , if nu then "NULL" else "NOT NULL"
     , case def of
         Nothing -> ""
-        Just s -> -- Avoid DEFAULT NULL, since it is always unnecessary, and is an error for text/blob fields
-                  if T.toUpper s == "NULL" then ""
-                  else " DEFAULT " ++ T.unpack s
+        Just s ->
+            -- Avoid DEFAULT NULL, since it is always unnecessary, and is an error for text/blob fields
+            if T.toUpper s == "NULL"
+                then ""
+                else " DEFAULT " ++ T.unpack s
     , case ref of
-        Nothing -> ""
-        Just cRef -> " REFERENCES " ++ escapeE (crTableName cRef)
-            <> " " <> T.unpack (renderFieldCascade (crFieldCascade cRef))
+        Just cRef | showReferences ->
+            mconcat
+                [ " REFERENCES "
+                , escapeE (crTableName cRef)
+                , " "
+                , T.unpack (renderFieldCascade (crFieldCascade cRef))
+                ]
+        _ ->
+            ""
     ]
 
 
@@ -1050,14 +1070,14 @@ showAlter table (Change (Column n nu t def gen defConstraintName maxLen _ref)) =
     , " CHANGE "
     , escapeF n
     , " "
-    , showColumn (Column n nu t def gen defConstraintName maxLen Nothing)
+    , showAlterColumn (Column n nu t def gen defConstraintName maxLen Nothing)
     ]
 showAlter table (Add' col) =
     concat
     [ "ALTER TABLE "
     , escapeE table
     , " ADD COLUMN "
-    , showColumn col
+    , showAlterColumn col
     ]
 showAlter table (Drop c) =
     concat